samba: pass down size_t instead of int to add_string_to_array().
Guenther


Autobuild-User(master): Günther Deschner <gd@samba.org>
Autobuild-Date(master): Mon Nov 17 19:53:22 CET 2014 on sn-devel-104

The last 3 patches address
BUG: https://bugzilla.samba.org/show_bug.cgi?id=10942
cleanup add_string_to_array and usage

+++ libcli/ldap/ldap_message.h
@@ -104,7 +104,7 @@ struct ldap_SearchRequest {
 	uint32_t sizelimit;
 	bool attributesonly;
 	struct ldb_parse_tree *tree;
-	int num_attributes;
+	size_t num_attributes;
 	const char * const *attributes;
 };
